backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
psycopg[binary]>=3.2.13,<3.4
psycopg-pool>=3.2.7,<3.4
python-dotenv==1.0.1
